refactor to properly use a workspace; add start and end dates to campaigns; prevent adding keys after a campaign is over

+++ near-campaign/src/clear_state.rs
@@ -12,9 +12,16 @@ impl Campaign {
   pub fn clear_state(&mut self, keys: Vec<PublicKey>) -> ClearStatus {
     keys.into_iter().for_each(|pk| {
       let key = pk.into();
-      self.keys.remove(&key);
-      // TODO only if status == 'active' remove access key
-      Promise::new(env::current_account_id()).delete_key(key);
+      let key_status = self.keys.remove(&key);
+
+      match key_status {
+        Some(status) => { 
+          if status == KeyStatus::Active {
+            Promise::new(env::current_account_id()).delete_key(key);
+          }
+        },
+        None => (),
+      };
     });
 
     ClearStatus::Completed(self.keys.is_empty())
